added dataClasses.cs
