+++ Videogame/LostIllusion/Assets/Scripts/dataClasses.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b146487f58885a94496fa614b90963ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
